+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 - (DEV) Fixes the `gulp README-to-local` and `gulp README-to-online`.
 - (DEV) Fixes code examples on JSDocs for `gulp build-docs`.
+### Changed
+- (DEV) `npm run test-all-serial` runs all the tests in serial.
 ### Added
 - `gulp build-docs` adds links for property and method classes.